citaa_system/requirements.txt
# CITAA System Requirements
# Flet Framework
flet>=0.21.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0>=0.21.0

# Google APIs
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
gspread>=5.11.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.2

# PDF Generation (for roster program)
reportlab>=4.0.0
PyPDF2>=3.0.0

# Optional: For better font support
# Pillow>=10.0.0
